move File handling into its own hook
